Add error messages for incomplete parsing
As result of applying

> runParser . prs

to a boomerang

> b :: Boomerang e tok a b

and initial position

> initalPosition (Nothing :: Maybe e)

one gets a list consisting of three type of elements:

- those that resulted from errors encoded in primitive parsers of the form

> Left e

where e contains error message and position where the error occur;

- those that produce complete successful parsing of the form

> Right ((f,tok), pos)

where tok is empty in an appropriate sense;

- and those where the parsing stopped before the full input whose exhausted of
  the form

> Right ((f,tok), pos)

with non-empty tok and pos been the first place starting from which no parsing
was made.

In the previous version the elements of last type were silently thrown away,
which created incomplete error messages. For example

> parseString Text.Boomerang.String.alpha "a("

was given the message

This patch fixes the above issue by explicitly handling the results of parsing
of the third kind in the function

> parse1

This led to the following change of API:

- the result of the function "parse" now includes an extra field that contains
  the position of the first non-parsed element;

- the class ErrorPosition got an extra method "setPosition" that permits to add
  position to the messages generated in "parse1" for incomplete parses;

- the instance of "ErrorPosition (ParserError p)" is changed according to the
  new interface of the class ErrorPosition.