Allow building with template-haskell-2.18.*
`template-haskell-2.18`, which is bundled with GHC 9.2, added a `[Type]`
field to `ConP` to represent visible type arguments. (See
https://gitlab.haskell.org/ghc/ghc/-/wikis/migration/9.2?version_id=ad6254b755131de4401bcc73afcb2e6e7542015a#template-haskell-218).
As a result, the uses of `ConP` in `Text.Boomerang.TH` need to be adjusted to
allow them to compile with GHC 9.2. I opted to factor out each use of `ConP`
into a `conPCompat` function in order to minimize the amount of CPP needed.